+++ src/test/java/ru/yoomoney/tech/plugins/grafana/public/PublicGrafanaDashboardPluginSpec.kt
@@ -1,7 +1,7 @@
-package ru.yandex.money.gradle.plugins.grafana.public
+package ru.yoomoney.tech.plugins.grafana.public
 
 import org.junit.Test
-import ru.yandex.money.gradle.plugins.grafana.GrafanaDashboardPluginSpec
+import ru.yoomoney.tech.plugins.grafana.GrafanaDashboardPluginSpec
 
 class PublicGrafanaDashboardPluginSpec : GrafanaDashboardPluginSpec() {
     override var repositories: String
@@ -12,7 +12,7 @@ class PublicGrafanaDashboardPluginSpec : GrafanaDashboardPluginSpec() {
         """.trimIndent()
         set(_) {}
     override var pluginId: String
-        get() = "com.yandex.money.tech.grafana-dashboard-plugin"
+        get() = "ru.yoomoney.tech.grafana-dashboard-plugin"
         set(_) {}
 
     @Test